Move the local logic to separate process
